refactor(all): renamed utils to util, use defaults, some docblocks
